1. I changed to setting that does not depend on Configuration.h. This is to make it easier to reuse modules. Specifically, we made it possible to pass information in PHKNetwork IP class constructor. Example. PHKNetworkIP networkIP(deviceName,devicePassword,deviceIdentity,controllerRecordsAddress); do {     networkIP.handleConnection(); } while (true);
2.
I made PHKControllerRecord.cpp from a function to a class.
This is to make it possible to pass the path saving vector <PHKKeyRecord> from outside by changing (1).

The PHKControllerRecord class delegates administration to the AccessorySet.
This is because AccessorySet is singleton so it is easy to use.
Originally, I want it under PHKNetworkIP, but since it is hard to modify, I delegated to AccessorySet for the time being.

3.
I changed have taken measures against buffer overrun of scanf.
Example.
We changed it as

char buf[10];
sccanf(a, "%s", buf);

 --->

char buf[10+1];
sccanf (a, "%10s", buf);

4.
Changed it to a code that is harder to leak memory.
Example.
We changed it as

char* a = new[100];

 --->

vector<char> a_vec(100); char* a = &a_vec[0];

If secured with vector<char>, it is safe because you do not have to delete it.

5.
I introduced AccessorySetAutoLock autolock; to avoid forgetting mutex lock release.
When you leave the scope, it automatically releases the mutex.
{
   lock

   ....

   unlock
}

---->

{
   AccessorySetAutoLock autolock

   ....
} //auto unlock

6.
I changed the method of initialization.
I use this as C ++ will initialize.
char a[100]; bzero(b,100);

----->

char a[100] = {0};
(It is very easy because C ++ will initialize without permission.)

7.
Fixed a problem that boolCharacteristics "1" can not be judged as true.

8.
Added PHKNetworkIP::closeAcceptConnection command to end accept loop.
Now you can stop with CTRL + C, and reusability as a module increases.

For details,plasess see the end of main.cpp.

PHKNetworkIP networkIP(deviceName,devicePassword,deviceIdentity,controllerRecordsAddress);
do {
    networkIP.handleConnection();
} while (true);

//if you running PHKNetworkIP new accept thread,
//call PHKNetworkIP::closeAcceptConnection to stop.

9.
To fix memory leak, we fixed to release memory when releasing Service and Characteristics.

10.
Uint8 Characteristic support added.
(However, perhaps, this may not be necessary ...?? )

11.
We have supported the number of digits up to 8 digits.
(Question: Why are you making three digits, why do you have [3] a reason?)

char buf[3];
snprintf(buf,3,"%d",intValue);

--->

char buf[8];
snprintf(buf, 8, "%d", intValue);

(I wrote this sentence using google translation. Sorry if we had funny English.)